Folder/file restructuring for .jsx and .css for respective components
File cleanup and restructuring for Gameplay, Cutscene, Menu, SelectMap, Profile, RoleAssignment and SelectMap in src-->pages
