+++ src/pages/CutScene/Cutscene.jsx
@@ -1,5 +1,6 @@
 import React, { useEffect } from "react";
 import { useNavigate } from 'react-router-dom';
+import '../../App.css';
 
 const Cutscene = () => {
   const navigate = useNavigate();
