+++ src/pages/GamePlay/Gameplay.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useParams, useNavigate } from 'react-router-dom';
-import '../App.css';
+import '../../App.css';
 
 const GamePlay = () => {
   const { role } = useParams();
