+++ src/pages/MainMenu/Menu.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useNavigate } from 'react-router-dom';
-import '../App.css';
-import '../index.css';
+import '../../App.css';
+import '../../index.css';
 
 const Menu = () => {
   const navigate = useNavigate();
