+++ src/pages/Profile/Profile.css
@@ -7,6 +7,8 @@
 .profile-container {
   display: flex;
   height: 80px;
+  flex-wrap: wrap;
+  justify-content: center;
   /* Full height */
   align-content: center;
   align-items: center;
@@ -70,8 +72,12 @@
   height: auto;
   /* Maintain aspect ratio */
   border-radius: 20px;
-  /* Makes the image circular */
+  transition: transform 0.3s ease;
   border: solid 5px rgba(36, 35, 35, 0.253);
+}
+
+.profile-img:hover {
+  transform: scale(1.05); /* Subtle hover effect */
 }
 
 .image-options {
@@ -114,6 +120,11 @@
 .submit-btn{
   padding: 10px;
   margin-top: 20px;
+  background-color: var(--primary-accent);
+  transition: background-color 0.3s ease;
+}
+.submit-btn:hover {
+  background-color: var(--secondary-accent);
 }
 
 .enter-name-form{
@@ -181,4 +192,22 @@
 
 .side-by-side-left-button{
   margin-left: 10px;
+}
+
+
+.enter-info {
+  background-color: #f9f9f9;
+  border-radius: 40px;
+  border: 1px solid transparent;
+  padding: 0.6em 1.2em;
+  font-size: 1em;
+  font-weight: 500;
+  font-family: inherit;
+  background-color: #f3f5efe0;
+  color: black;
+  margin: 10px;
+  transition: border-color 0.25s;
+  margin-top: 10px;
+
+
 }
